status var changed from Short to Status (enum) on findAllByUserId(...)

+++ CentralDeErros/src/main/java/com/squad2/CentralDeErros/repository/LogRepository.java
@@ -21,6 +21,6 @@ public interface LogRepository extends JpaRepository<Log, Long> {
             "WHERE ID_USUARIO_ID_USUARIO = :userId " +
             "AND STATUS = :status",
             nativeQuery = true)
-    List<Log> findAllByUserId(@Param("userId") Long userId, @Param("status") Short status, Pageable pageable);
+    List<Log> findAllByUserId(@Param("userId") Long userId, @Param("status") int status, Pageable pageable);
 
 }
